requirements.txt
streamlit
pandas
torch --index-url https://download.pytorch.org/whl/cpu
sentence-transformers
Pillow